opt: 优化coord session，与pmdRemoteSession代码合一
git-svn-id: http://192.168.20.11/sequoiadb/trunk@28907 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/coord/coordResource.cpp
@@ -1775,7 +1775,7 @@ namespace engine
          }
       }
 
-      SDB_ASSERT( objNum == pReply->numReturned,
+      SDB_ASSERT( objNum == (UINT32)pReply->numReturned,
                   "Return number is invalid" ) ;
 
       if ( objNum < 1 || isSpec )
